+++ h2integrate/converters/methanol/smr_methanol_plant.py
@@ -1,5 +1,6 @@
 import numpy as np
 from attrs import field, define
+from openmdao.utils.units import convert_units
 
 from h2integrate.core.utilities import merge_shared_inputs
 from h2integrate.converters.methanol.methanol_baseclass import (
@@ -16,27 +17,27 @@ from h2integrate.converters.methanol.methanol_baseclass import (
 class SMRPerformanceConfig(MethanolPerformanceConfig):
     meoh_syn_cat_consume_ratio: float = field()
     meoh_atr_cat_consume_ratio: float = field()
-    lng_consume_ratio: float = field()
+    ng_consume_ratio: float = field()
     elec_produce_ratio: float = field()
 
 
 class SMRMethanolPlantPerformanceModel(MethanolPerformanceBaseClass):
     """
-    An OpenMDAO component for modeling the performance of a methanol plant.
-    Computes annual methanol and co-product production, feedstock consumption, and emissions
+    An OpenMDAO component for modeling the performance of a steam methane reforming (SMR) methanol
+    plant. Computes annual methanol and co-product production, feedstock consumption, and emissions
     based on plant capacity and capacity factor.
 
     Inputs:
-        - meoh_syn_cat_consume_ratio: ratio of methanol synthesis catalyst consumed to
+        - meoh_syn_cat_consume_ratio: ratio of ft^3 methanol synthesis catalyst consumed to
             kg methanol produced
-        - meoh_atr_cat_consume_ratio: ratio of methanol ATR catalyst consumed to
-            kg methanol produced
-        - lng_consume_ratio: ratio of LNG consumed to kg methanol produced
+        - meoh_atr_cat_consume_ratio: ratio of ft^3 methanol autothermal reforming (ATR) catalyst
+            consumed to kg methanol produced
+        - ng_consume_ratio: ratio of kg natural gas (NG) consumed to kg methanol produced
         - elec_produce_ratio: ratio of electricity produced to kg methanol produced
     Outputs:
         - meoh_syn_cat_consumption: annual consumption of methanol synthesis catalyst (ft**3/yr)
         - meoh_atr_cat_consumption: annual consumption of methanol ATR catalyst (ft**3/yr)
-        - lng_consumption: hourly consumption of LNG (kg/h)
+        - ng_consumption: hourly consumption of NG (kg/h)
         - electricity: hourly electricity consumption (kW*h/h)
     """
 
@@ -57,12 +58,12 @@ class SMRMethanolPlantPerformanceModel(MethanolPerformanceBaseClass):
             units="ft**3/kg",
             val=self.config.meoh_atr_cat_consume_ratio,
         )
-        self.add_input("lng_consume_ratio", units="kg/kg", val=self.config.lng_consume_ratio)
+        self.add_input("ng_consume_ratio", units="kg/kg", val=self.config.ng_consume_ratio)
         self.add_input("elec_produce_ratio", units="kW*h/kg", val=self.config.elec_produce_ratio)
 
         self.add_output("meoh_syn_cat_consumption", units="ft**3/yr")
         self.add_output("meoh_atr_cat_consumption", units="ft**3/yr")
-        self.add_output("lng_consumption", shape=8760, units="kg/h")
+        self.add_output("ng_consumption", shape=8760, units="kg/h")
         self.add_output("electricity", shape=8760, units="kW*h/h")
 
     def compute(self, inputs, outputs):
@@ -87,35 +88,35 @@ class SMRMethanolPlantPerformanceModel(MethanolPerformanceBaseClass):
         outputs["meoh_syn_cat_consumption"] = (
             np.sum(meoh_kgph) * inputs["meoh_syn_cat_consume_ratio"]
         )
-        outputs["lng_consumption"] = meoh_kgph * inputs["lng_consume_ratio"]
+        outputs["ng_consumption"] = meoh_kgph * inputs["ng_consume_ratio"]
         outputs["electricity"] = meoh_kgph * inputs["elec_produce_ratio"]
 
 
 @define
 class SMRCostConfig(MethanolCostConfig):
+    ng_lhv: float = field()
     meoh_syn_cat_price: float = field()
     meoh_atr_cat_price: float = field()
-    lng_price: float = field()
-    elec_sales_price: float = field()
+    ng_price: float = field()
 
 
 class SMRMethanolPlantCostModel(MethanolCostBaseClass):
     """
-    An OpenMDAO component for modeling the cost of an SMR methanol plant.
+    An OpenMDAO component for modeling the cost of a steam methane (SMR) reforming methanol plant.
 
     Inputs:
+        ng_lhv: natural gas lower heating value in MJ/kg
         meoh_syn_cat_consumption: annual consumption of methanol synthesis catalyst (ft**3/yr)
         meoh_atr_cat_consumption: annual consumption of methanol ATR catalyst (ft**3/yr)
-        lng_consumption: hourly consumption of LNG (kg/h)
+        ng_consumption: hourly consumption of NG (kg/h)
         electricity: hourly electricity consumption (kW*h/h)
         meoh_syn_cat_price: price of methanol synthesis catalyst (USD/ft**3)
         meoh_atr_cat_price: price of methanol ATR catalyst (USD/ft**3)
-        lng_price: price of LNG (USD/MBtu)
-        elec_sales_price: electricity sales price (USD/kW/h)
+        ng_price: price of NG (USD/MBtu)
     Outputs:
         meoh_syn_cat_cost: annual cost of methanol synthesis catalyst (USD/year)
         meoh_atr_cat_cost: annual cost of methanol ATR catalyst (USD/year)
-        lng_cost: annual cost of LNG (USD/year)
+        ng_cost: annual cost of NG (USD/year)
         elec_revenue: annual revenue from electricity sales (USD/year)
     """
 
@@ -125,29 +126,31 @@ class SMRMethanolPlantCostModel(MethanolCostBaseClass):
         )
         super().setup()
 
+        self.add_input("ng_lhv", units="MJ/kg", val=self.config.ng_lhv)
         self.add_input("meoh_syn_cat_consumption", units="ft**3/yr")
         self.add_input("meoh_atr_cat_consumption", units="ft**3/yr")
-        self.add_input("lng_consumption", shape=8760, units="kg/h")
+        self.add_input("ng_consumption", shape=8760, units="kg/h")
         self.add_input("electricity", shape=8760, units="kW*h/h")
         self.add_input("meoh_syn_cat_price", units="USD/ft**3", val=self.config.meoh_syn_cat_price)
         self.add_input("meoh_atr_cat_price", units="USD/ft**3", val=self.config.meoh_atr_cat_price)
         self.add_input(
-            "lng_price", units="USD/MBtu", val=self.config.lng_price
+            "ng_price", units="USD/MBtu", val=self.config.ng_price
         )  # TODO: get OpenMDAO to recognize 'MMBtu'
-        self.add_input("elec_sales_price", units="USD/kW/h", val=self.config.elec_sales_price)
 
         self.add_output("meoh_syn_cat_cost", units="USD/year")
         self.add_output("meoh_atr_cat_cost", units="USD/year")
-        self.add_output("lng_cost", units="USD/year")
+        self.add_output("ng_cost", units="USD/year")
         self.add_output("elec_revenue", units="USD/year")
 
     def compute(self, inputs, outputs):
-        MMBTU_per_GJ = 1.055
-        LHV = 0.0201  # GJ/kg
-
         toc_usd = inputs["plant_capacity_kgpy"] * inputs["toc_kg_y"]
         foc_usd_y = inputs["plant_capacity_kgpy"] * inputs["foc_kg_y2"]
         voc_usd_y = np.sum(inputs["methanol"]) * inputs["voc_kg"]
+
+        ppa_price = self.options["plant_config"]["plant"]["ppa_price"]
+
+        lhv_mj = inputs["ng_lhv"]
+        lhv_mmbtu = convert_units(lhv_mj, "MJ", "MBtu")
 
         outputs["CapEx"] = toc_usd
         outputs["OpEx"] = foc_usd_y + voc_usd_y
@@ -159,20 +162,19 @@ class SMRMethanolPlantCostModel(MethanolCostBaseClass):
         outputs["meoh_atr_cat_cost"] = (
             inputs["meoh_atr_cat_consumption"] * inputs["meoh_atr_cat_price"]
         )
-        outputs["lng_cost"] = (
-            np.sum(inputs["lng_consumption"]) * MMBTU_per_GJ * LHV * inputs["lng_price"]
-        )
-        outputs["elec_revenue"] = np.sum(inputs["electricity"]) * inputs["elec_sales_price"]
+        outputs["ng_cost"] = np.sum(inputs["ng_consumption"]) * lhv_mmbtu * inputs["ng_price"]
+        outputs["elec_revenue"] = np.sum(inputs["electricity"]) * ppa_price
 
 
 class SMRMethanolPlantFinanceModel(MethanolFinanceBaseClass):
     """
-    An OpenMDAO component for modeling the financing of a methanol plant.
+    An OpenMDAO component for modeling the financing of a steam methane reforming (SMR) methanol
+    plant.
 
     Inputs:
         meoh_syn_cat_cost: annual cost of synthesis catalyst in USD/year
         meoh_atr_cat_cost: annual cost of ATR catalyst in USD/year
-        lng_cost: annual cost of natural gas in USD/year
+        ng_cost: annual cost of natural gas in USD/year
         elec_revenue: annual revenue from electricity sales in USD/year
     Outputs:
         LCOM_meoh_atr_cat: levelized cost of methanol from ATR catalyst in USD/kg
@@ -198,7 +200,7 @@ class SMRMethanolPlantFinanceModel(MethanolFinanceBaseClass):
             desc="Annual cost of ATR catalyst in USD/year",
         )
         self.add_input(
-            "lng_cost",
+            "ng_cost",
             units="USD/year",
             desc="Annual cost of natural gas in USD/year",
         )
@@ -253,7 +255,7 @@ class SMRMethanolPlantFinanceModel(MethanolFinanceBaseClass):
         lcom_vopex -= lcom_meoh_syn_cat + lcom_meoh_atr_cat
         outputs["LCOM_meoh_vopex"] = lcom_vopex
 
-        ng_cost = inputs["lng_cost"]
+        ng_cost = inputs["ng_cost"]
         lcom_ng = ng_cost / np.sum(kgph)
         outputs["LCOM_ng"] = lcom_ng